refactor(scim): move enterpriseSCIMMetaSchema() to util_enterprise_scim.go
Pure refactor — no functional change. The shared schema helper was
defined in data_source_github_enterprise_scim_groups.go but referenced
by both groups and users data sources. Moving it to the shared util
file makes the codebase easier to navigate.

Closes #54
